database - updated to bigint

+++ database/create_skynet_db.py
@@ -1,7 +1,7 @@
 from sqlalchemy import create_engine, text, exc, DateTime
 from sqlalchemy import inspect
 from sqlalchemy.orm import declarative_base
-from sqlalchemy import Column, String, Integer, Float, Boolean, Date
+from sqlalchemy import Column, String, Integer, Float, Boolean, Date, BIGINT
 from datetime import datetime, timezone
 import os
 
@@ -113,7 +113,7 @@ class Aircraft(Base):
     spi = Column(Boolean)  # Special position identification (indicates an emergency or priority status).
     mlat = Column(String)  # Multilateration data (lists if position was derived from multilateration).
     tisb = Column(String)  # Traffic Information Service Broadcast (indicates if data comes from TIS-B).
-    messages = Column(Integer)  # Number of ADS-B messages received.
+    messages = Column(BIGINT)  # Number of ADS-B messages received.
     seen = Column(Float)  # Time since last update received.
     rssi = Column(Float)  # Received signal strength indicator.
     distance = Column(Float)  # Distance from the receiver (kilometers).
